ServerSideRender: Extract data fetching logic into a custom hook
