+++ packages/server-side-render/src/server-side-render.js
@@ -1,56 +1,14 @@
 /**
- * External dependencies
- */
-import fastDeepEqual from 'fast-deep-equal/es6';
-
-/**
  * WordPress dependencies
  */
-import { useDebounce, usePrevious } from '@wordpress/compose';
-import {
-	RawHTML,
-	useCallback,
-	useEffect,
-	useLayoutEffect,
-	useRef,
-	useState,
-} from '@wordpress/element';
+import { RawHTML, useEffect, useState } from '@wordpress/element';
 import { __, sprintf } from '@wordpress/i18n';
-import apiFetch from '@wordpress/api-fetch';
-import { addQueryArgs } from '@wordpress/url';
 import { Placeholder, Spinner } from '@wordpress/components';
-import { __experimentalSanitizeBlockAttributes } from '@wordpress/blocks';
 
-const EMPTY_OBJECT = {};
-
-export function rendererPath( block, attributes = null, urlQueryArgs = {} ) {
-	return addQueryArgs( `/wp/v2/block-renderer/${ block }`, {
-		context: 'edit',
-		...( null !== attributes ? { attributes } : {} ),
-		...urlQueryArgs,
-	} );
-}
-
-export function removeBlockSupportAttributes( attributes ) {
-	const {
-		backgroundColor,
-		borderColor,
-		fontFamily,
-		fontSize,
-		gradient,
-		textColor,
-		className,
-		...restAttributes
-	} = attributes;
-
-	const { border, color, elements, spacing, typography, ...restStyles } =
-		attributes?.style || EMPTY_OBJECT;
-
-	return {
-		...restAttributes,
-		style: restStyles,
-	};
-}
+/**
+ * Internal dependencies
+ */
+import { useServerSideRender } from './hook';
 
 function DefaultEmptyResponsePlaceholder( { className } ) {
 	return (
@@ -108,113 +66,10 @@ export default function ServerSideRender( props ) {
 		EmptyResponsePlaceholder = DefaultEmptyResponsePlaceholder,
 		ErrorResponsePlaceholder = DefaultErrorResponsePlaceholder,
 		LoadingResponsePlaceholder = DefaultLoadingResponsePlaceholder,
+		...restProps
 	} = props;
 
-	const isMountedRef = useRef( false );
-	const fetchRequestRef = useRef();
-	const [ response, setResponse ] = useState( null );
-	const prevProps = usePrevious( props );
-	const [ isLoading, setIsLoading ] = useState( false );
-	const latestPropsRef = useRef( props );
-
-	useLayoutEffect( () => {
-		latestPropsRef.current = props;
-	}, [ props ] );
-
-	const fetchData = useCallback( () => {
-		if ( ! isMountedRef.current ) {
-			return;
-		}
-
-		const {
-			attributes,
-			block,
-			skipBlockSupportAttributes = false,
-			httpMethod = 'GET',
-			urlQueryArgs,
-		} = latestPropsRef.current;
-
-		setIsLoading( true );
-
-		let sanitizedAttributes =
-			attributes &&
-			__experimentalSanitizeBlockAttributes( block, attributes );
-
-		if ( skipBlockSupportAttributes ) {
-			sanitizedAttributes =
-				removeBlockSupportAttributes( sanitizedAttributes );
-		}
-
-		// If httpMethod is 'POST', send the attributes in the request body instead of the URL.
-		// This allows sending a larger attributes object than in a GET request, where the attributes are in the URL.
-		const isPostRequest = 'POST' === httpMethod;
-		const urlAttributes = isPostRequest
-			? null
-			: sanitizedAttributes ?? null;
-		const path = rendererPath( block, urlAttributes, urlQueryArgs );
-		const data = isPostRequest
-			? { attributes: sanitizedAttributes ?? null }
-			: null;
-
-		// Store the latest fetch request so that when we process it, we can
-		// check if it is the current request, to avoid race conditions on slow networks.
-		const fetchRequest = ( fetchRequestRef.current = apiFetch( {
-			path,
-			data,
-			method: isPostRequest ? 'POST' : 'GET',
-		} )
-			.then( ( fetchResponse ) => {
-				if (
-					isMountedRef.current &&
-					fetchRequest === fetchRequestRef.current &&
-					fetchResponse
-				) {
-					setResponse( fetchResponse.rendered );
-				}
-			} )
-			.catch( ( error ) => {
-				if (
-					isMountedRef.current &&
-					fetchRequest === fetchRequestRef.current
-				) {
-					setResponse( {
-						error: true,
-						errorMsg: error.message,
-					} );
-				}
-			} )
-			.finally( () => {
-				if (
-					isMountedRef.current &&
-					fetchRequest === fetchRequestRef.current
-				) {
-					setIsLoading( false );
-				}
-			} ) );
-
-		return fetchRequest;
-	}, [] );
-
-	const debouncedFetchData = useDebounce( fetchData, 500 );
-
-	// When the component unmounts, set isMountedRef to false. This will
-	// let the async fetch callbacks know when to stop.
-	useEffect( () => {
-		isMountedRef.current = true;
-		return () => {
-			isMountedRef.current = false;
-		};
-	}, [] );
-
-	useEffect( () => {
-		// Don't debounce the first fetch. This ensures that the first render
-		// shows data as soon as possible.
-		if ( prevProps === undefined ) {
-			fetchData();
-		} else if ( ! fastDeepEqual( prevProps, props ) ) {
-			debouncedFetchData();
-		}
-	} );
+	const { response, isLoading } = useServerSideRender( restProps );
 
 	const hasResponse = !! response;
 	const hasEmptyResponse = response === '';
