+++ packages/server-side-render/src/test/index.js
@@ -1,10 +1,7 @@
 /**
  * Internal dependencies
  */
-import {
-	rendererPath,
-	removeBlockSupportAttributes,
-} from '../server-side-render';
+import { rendererPath, removeBlockSupportAttributes } from '../hook';
 
 describe( 'rendererPath', () => {
 	test( 'should return an base path for empty input', () => {
